requirements.txt
flask
pyngrok
pytube
moviepy
pyffmpeg
pydub
langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
textract
openai
tiktoken
faiss-cpu